Reorganised IO functions
IO functions are now in the 'io.py' script instead of in the datastore utils. Preparation for expanding IO for other input file types.
